+++ src/main/resources/create_table_students.sql
@@ -4,8 +4,8 @@
 
 CREATE TABLE public.students
 (
-    student_id bigint NOT NULL DEFAULT nextval('students_student_id_seq'::regclass),
-    group_id bigint,
+    student_id integer NOT NULL DEFAULT nextval('students_student_id_seq'::regclass),
+    group_id integer,
     first_name character varying(50) COLLATE pg_catalog."default" NOT NULL,
     last_name character varying(50) COLLATE pg_catalog."default" NOT NULL,
     CONSTRAINT students_pkey PRIMARY KEY (student_id),
